+++ src/game/TileEngine/Map_Edgepoints.cc
@@ -1018,7 +1018,7 @@ void ChooseMapEdgepoints(MAPEDGEPOINTINFO* const pMapEdgepointInfo, const UINT8 
 	if (ubNumDesiredPoints >= n_usable)
 	{ //We don't have enough points for everyone, return them all.
 		pMapEdgepointInfo->ubNumPoints = static_cast<UINT8>(n_usable);
-		for (INT32 i = 0; i < n_usable; ++i)
+		for (size_t i = 0; i < n_usable; ++i)
 		{
 			pMapEdgepointInfo->sGridNo[i] = psTempArray[i];
 		}
@@ -1029,7 +1029,7 @@ void ChooseMapEdgepoints(MAPEDGEPOINTINFO* const pMapEdgepointInfo, const UINT8 
 	UINT16 usSlots    = static_cast<UINT16>(n_usable);
 	UINT16 usCurrSlot = 0;
 	pMapEdgepointInfo->ubNumPoints = ubNumDesiredPoints;
-	for (INT32 i = 0; i < n_usable; ++i)
+	for (size_t i = 0; i < n_usable; ++i)
 	{
 		if (Random(usSlots) < ubNumDesiredPoints)
 		{
