+++ src/game/TileEngine/Radar_Screen.cc
@@ -440,13 +440,13 @@ static void RenderSquadList(void)
 	SetFont(SQUAD_FONT);
 	SetFontBackground(FONT_BLACK);
 
-	for (UINT i = 0; i < NUMBER_OF_SQUADS; ++i)
+	for (INT16 i = 0; i < NUMBER_OF_SQUADS; ++i)
 	{
 		const UINT8 colour =
 			sSelectedSquadLine == i         ? FONT_WHITE   : // highlight line?
 			!IsSquadOnCurrentTacticalMap(i) ? FONT_BLACK   :
-			CurrentSquad() == (INT32)i      ? FONT_LTGREEN :
-								FONT_DKGREEN;
+			CurrentSquad() == i             ? FONT_LTGREEN :
+			                                  FONT_DKGREEN;
 		SetFontForeground(colour);
 
 		INT16 sX;
